virtcontainers: cgroups: Move non Linux routine to utils.go
Have an OS agnostic file for sharing routines.
